fix(frontend): rename path for-builders and for-lawyers to web-three and enterprise
