+++ frontend/src/app/enterprise/components/Card.tsx
