+++ frontend/src/app/enterprise/components/Hero.tsx
