+++ frontend/src/app/enterprise/components/KlerosEnterpriseSection/HighlightedText.tsx
