+++ frontend/src/app/enterprise/components/KlerosEnterpriseSection/MethodsTable/DesktopTable.tsx
