+++ frontend/src/app/enterprise/components/KlerosEnterpriseSection/index.tsx
