+++ frontend/src/app/enterprise/queries/hero.ts
